Remove MobileContext and use PreferencesContext instead
